Comments from Discord per pepper
