+++ fabric-registry-sync-v0/src/main/java/net/fabricmc/fabric/impl/registry/sync/FabricRegistryInit.java
@@ -16,6 +16,8 @@
 
 package net.fabricmc.fabric.impl.registry.sync;
 
+import net.fabricmc.fabric.impl.registry.sync.registryentrylists.defaults.DefaultCustomRegistryEntryListsImpl;
+
 import net.minecraft.registry.Registries;
 
 import net.fabricmc.api.ModInitializer;
@@ -25,7 +27,6 @@ import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
 import net.fabricmc.fabric.api.networking.v1.ServerConfigurationConnectionEvents;
 import net.fabricmc.fabric.api.networking.v1.ServerConfigurationNetworking;
 import net.fabricmc.fabric.impl.registry.sync.packet.DirectRegistryPacketHandler;
-import net.fabricmc.fabric.impl.registry.sync.registryentrylists.CustomRegistryEntryListImpl;
 import net.fabricmc.fabric.impl.registry.sync.registryentrylists.util.ChannelUtil;
 
 public class FabricRegistryInit implements ModInitializer {
@@ -215,6 +216,6 @@ public class FabricRegistryInit implements ModInitializer {
 				.addAttribute(RegistryAttribute.SYNCED);
 
 		ChannelUtil.init();
-		CustomRegistryEntryListImpl.init();
+		DefaultCustomRegistryEntryListsImpl.register();
 	}
 }
