--- fabric-registry-sync-v0/src/main/java/net/fabricmc/fabric/impl/registry/sync/registryentrylists/defaults/DefaultRegistryEntryLists.java
@@ -1,35 +1,0 @@
-/*
- * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package net.fabricmc.fabric.impl.registry.sync.registryentrylists.defaults;
-
-import net.fabricmc.fabric.impl.registry.sync.registryentrylists.CustomRegistryEntryListImpl;
-
-public class DefaultRegistryEntryLists {
-	private static boolean initialized = false;
-
-	public static void registerDefaults() {
-		if (initialized) {
-			return;
-		}
-
-		initialized = true;
-		CustomRegistryEntryListImpl.registerSerializer(UnionRegistryEntryList.SERIALIZER);
-		CustomRegistryEntryListImpl.registerSerializer(IntersectionRegistryEntryList.SERIALIZER);
-		CustomRegistryEntryListImpl.registerSerializer(InverseRegistryEntryList.SERIALIZER);
-		CustomRegistryEntryListImpl.registerSerializer(UniversalRegistryEntryList.SERIALIZER);
-	}
-}
