+++ fabric-registry-sync-v0/src/main/java/net/fabricmc/fabric/impl/registry/sync/registryentrylists/defaults/UnionRegistryEntryList.java
@@ -22,6 +22,7 @@ import java.util.stream.Collectors;
 
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
+import net.fabricmc.fabric.api.event.registry.CustomRegistryEntryListSerializer;
 import org.jetbrains.annotations.NotNull;
 
 import net.minecraft.network.RegistryByteBuf;
@@ -37,7 +38,7 @@ import net.minecraft.util.Identifier;
 import net.fabricmc.fabric.api.event.registry.CustomRegistryEntryList;
 
 public class UnionRegistryEntryList<T> extends MultiPartRegistryEntryList<T> {
-	public static final CustomRegistryEntryList.CustomRegistryEntryListSerializer SERIALIZER = new Serializer();
+	public static final CustomRegistryEntryListSerializer SERIALIZER = new Serializer();
 
 	public UnionRegistryEntryList(List<RegistryEntryList<T>> parts) {
 		super(parts);
